UserController
Now, let's create UserController that will process the login request.
Once a user is logged in successfully, we would like to keep this
information in the session. The presence of this object in the session
can be used to check whether the user is already logged in. Create the
UserController class in the packt.jee.course_management.  controller
package.