@RequestMapping (value="/login", method=RequestMethod.POST)
Note that we have specified the method attribute in the @RequestMapping
annotation. When the request URL is /login and the HTTP request type is
GET, only then will the login method be called. This method would not be
called if the POST request is sent from the client. In the login method,
we create an instance of UserDTO and add it to Model so that it is
accessible to view.  We will add a method to handle the POST request
from the login page. We will keep the same URL, that is, /login