@SessionAttributes("user")
We now have two methods in UserController handling the request URL
/login.  However, the login method handles the GET request and doLogin
handles the POST request. If authentication is successful in the doLogin
method, then we redirect to the courses (list) page. Else, we set an
error message and return to the login page.  We want to save the user
object created in the login method in the HTTP session.  This can be
done with a simple @SessionAttributes annotation. You can specify the
list of attributes in Model that need to be saved in the session too.
Further, we want to save the user attribute of Model in the session.
Therefore, we will add the following annotation to the UserController
class: