interceptor servlet-context.xml
The page expects user (instance of UserDTO) to be available. This is
made available by UserController through Model.  We now have a login
page and UserController to handle the authentication, but how do we make
sure this page is displayed for every request when the user is not
already logged in? This is where we can use a Spring interceptor. We
will configure the interceptor in the Spring context configuration file:
servlet-context.xml. Add the following code to servlet-context.xml: